refactor(pw): move route and profile into pw/types.h
